+++ package.json
@@ -10,9 +10,7 @@
   },
   "dependencies": {
     "@heroicons/react": "^2.1.5",
-    "autoprefixer": "^10.4.20",
-    "next": "^14.2.14",
-    "postcss-cli": "^11.0.0",
+    "next": "14.2.8",
     "react": "^18",
     "react-dom": "^18"
   },
