+++ src/app/auth/signup/upload-done/page.tsx
@@ -8,7 +8,7 @@ const VerificationPage: React.FC = () => {
       {/* Main container */}
       <div className="flex flex-col items-center gap-[64px] w-[918px]">
         {/* Title */}
-        <h1 className="text-header text-center">
+        <h1 className="text-[32px] font-semibold text-center leading-[120%]">
           We’ve received your request and are currently verifying the
           authenticity of your documents
         </h1>
@@ -18,20 +18,19 @@ const VerificationPage: React.FC = () => {
           <Image
             src="/Check your inbox.svg" // Make sure the path is correct
             alt="Document Verification"
-            width={180}
-            height={180}
+            width={314}
+            height={252}
           />
         </div>
 
-        {/* Message */}
-        <p className="text-subheading2 text-dark-blue">
-          You will hear back from us within 7 days
-        </p>
-
-        {/* Back to homepage link */}
-        <Link href="/dashboard/home" className="text-button text-soft-gray">
-          Back to homepage
-        </Link>
+        <div className="flex flex-col items-center gap-8">
+          <p className="text-subheading2 text-dark-blue">
+            You will hear back from us within 7 days
+          </p>
+          <Link href="/dashboard/home" className="text-body-bold text-soft-gray">
+            Back to homepage
+          </Link>
+        </div>
       </div>
     </div>
   );
